journal, favorites still not rendering
