+++ journal.html
@@ -125,6 +125,6 @@
 </div>
 </div>
 </div>
-<script type="module" src="./src/Paul/app.js"></script>
+<script type="module" src="./src/IlonaPaul/app.js"></script>
   </body>
 </html>
